Refactor function_die.rs from debug/mod.rs
