+++ probe-rs/src/debug/mod.rs
@@ -10,6 +10,8 @@
 
 /// Debug information which is parsed from DWARF debugging information.
 pub mod debug_info;
+/// References to the DIE (debug information entry) of functions.
+pub mod function_die;
 /// Target Register definitions.
 pub mod registers;
 /// The stack frame information used while unwinding the stack from a specific program counter.
@@ -112,159 +114,8 @@ pub struct SourceLocation {
     pub high_pc: Option<u32>,
 }
 
-type FunctionDieType<'abbrev, 'unit> =
-    gimli::DebuggingInformationEntry<'abbrev, 'unit, debug_info::GimliReader, usize>;
-
 type UnitIter =
     gimli::DebugInfoUnitHeadersIter<gimli::EndianReader<gimli::LittleEndian, std::rc::Rc<[u8]>>>;
-
-/// Reference to a DIE for a function
-pub(crate) struct FunctionDie<'abbrev, 'unit, 'unit_info, 'debug_info> {
-    unit_info: &'unit_info UnitInfo<'debug_info>,
-
-    function_die: FunctionDieType<'abbrev, 'unit>,
-
-    /// Only present for inlined functions, where this is a reference
-    /// to the declaration of the function.
-    abstract_die: Option<FunctionDieType<'abbrev, 'unit>>,
-    /// The address of the first instruction in this function.
-    low_pc: u64,
-    /// The address of the first instruction after this funciton.
-    high_pc: u64,
-}
-
-impl<'debugunit, 'abbrev, 'unit: 'debugunit, 'unit_info, 'debug_info>
-    FunctionDie<'abbrev, 'unit, 'unit_info, 'debug_info>
-{
-    fn new(
-        die: FunctionDieType<'abbrev, 'unit>,
-        unit_info: &'unit_info UnitInfo<'debug_info>,
-    ) -> Option<Self> {
-        let tag = die.tag();
-
-        match tag {
-            gimli::DW_TAG_subprogram => Some(Self {
-                unit_info,
-                function_die: die,
-                abstract_die: None,
-                low_pc: 0,
-                high_pc: 0,
-            }),
-            other_tag => {
-                log::error!("FunctionDie has to has to have Tag DW_TAG_subprogram, but tag is {:?}. This is a bug, please report it.", other_tag.static_string());
-                None
-            }
-        }
-    }
-
-    fn new_inlined(
-        concrete_die: FunctionDieType<'abbrev, 'unit>,
-        abstract_die: FunctionDieType<'abbrev, 'unit>,
-        unit_info: &'unit_info UnitInfo<'debug_info>,
-    ) -> Option<Self> {
-        let tag = concrete_die.tag();
-
-        match tag {
-            gimli::DW_TAG_inlined_subroutine => Some(Self {
-                unit_info,
-                function_die: concrete_die,
-                abstract_die: Some(abstract_die),
-                low_pc: 0,
-                high_pc: 0,
-            }),
-            other_tag => {
-                log::error!("FunctionDie has to has to have Tag DW_TAG_inlined_subroutine, but tag is {:?}. This is a bug, please report it.", other_tag.static_string());
-                None
-            }
-        }
-    }
-
-    fn is_inline(&self) -> bool {
-        self.abstract_die.is_some()
-    }
-
-    fn function_name(&self) -> Option<String> {
-        if let Some(fn_name_attr) = self.get_attribute(gimli::DW_AT_name) {
-            match fn_name_attr.value() {
-                gimli::AttributeValue::DebugStrRef(fn_name_ref) => {
-                    match self.unit_info.debug_info.dwarf.string(fn_name_ref) {
-                        Ok(fn_name_raw) => Some(String::from_utf8_lossy(&fn_name_raw).to_string()),
-                        Err(error) => {
-                            log::debug!("No value for DW_AT_name: {:?}: error", error);
-
-                            None
-                        }
-                    }
-                }
-                value => {
-                    log::debug!("Unexpected attribute value for DW_AT_name: {:?}", value);
-                    None
-                }
-            }
-        } else {
-            log::debug!("DW_AT_name attribute not found, unable to retrieve function name");
-            None
-        }
-    }
-
-    /// Get the call site of an inlined function.
-    ///
-    /// If this function is not inlined (`is_inline()` returns false),
-    /// this function returns `None`.
-    fn inline_call_location(&self) -> Option<SourceLocation> {
-        if !self.is_inline() {
-            return None;
-        }
-
-        let file_name_attr = self.get_attribute(gimli::DW_AT_call_file)?;
-
-        let (directory, file) = extract_file(
-            self.unit_info.debug_info,
-            &self.unit_info.unit,
-            file_name_attr.value(),
-        )?;
-        let line = self
-            .get_attribute(gimli::DW_AT_call_line)
-            .and_then(|line| line.udata_value());
-
-        let column =
-            self.get_attribute(gimli::DW_AT_call_column)
-                .map(|column| match column.udata_value() {
-                    None => ColumnType::LeftEdge,
-                    Some(c) => ColumnType::Column(c),
-                });
-        Some(SourceLocation {
-            line,
-            column,
-            file: Some(file),
-            directory: Some(directory),
-            low_pc: Some(self.low_pc as u32),
-            high_pc: Some(self.high_pc as u32),
-        })
-    }
-
-    /// Resolve an attribute by looking through both the origin or abstract die entries.
-    fn get_attribute(&self, attribute_name: gimli::DwAt) -> Option<debug_info::GimliAttribute> {
-        let attribute = self
-            .function_die
-            .attr(attribute_name)
-            .map_or(None, |attribute| attribute);
-
-        // For inlined function, the *abstract instance* has to be checked if we cannot find the
-        // attribute on the *concrete instance*.
-        if self.is_inline() && attribute.is_none() {
-            if let Some(origin) = self.abstract_die.as_ref() {
-                origin
-                    .attr(attribute_name)
-                    .map_or(None, |attribute| attribute)
-            } else {
-                None
-            }
-        } else {
-            attribute
-        }
-    }
-}
 
 pub(crate) struct UnitInfo<'debuginfo> {
     debug_info: &'debuginfo debug_info::DebugInfo,
@@ -279,7 +130,7 @@ impl<'debuginfo> UnitInfo<'debuginfo> {
         &self,
         address: u64,
         find_inlined: bool,
-    ) -> Result<Vec<FunctionDie>, DebugError> {
+    ) -> Result<Vec<function_die::FunctionDie>, DebugError> {
         log::trace!("Searching Function DIE for address {:#010x}", address);
 
         let mut entries_cursor = self.unit.entries();
@@ -292,7 +143,8 @@ impl<'debuginfo> UnitInfo<'debuginfo> {
                     if (ranges.begin <= address) && (address < ranges.end) {
                         // Check if we are actually in an inlined function
 
-                        if let Some(mut die) = FunctionDie::new(current.clone(), self) {
+                        if let Some(mut die) = function_die::FunctionDie::new(current.clone(), self)
+                        {
                             die.low_pc = ranges.begin;
                             die.high_pc = ranges.end;
 
@@ -337,7 +189,7 @@ impl<'debuginfo> UnitInfo<'debuginfo> {
         &self,
         address: u64,
         offset: UnitOffset,
-    ) -> Result<Vec<FunctionDie>, DebugError> {
+    ) -> Result<Vec<function_die::FunctionDie>, DebugError> {
         let mut current_depth = 0;
 
         let mut abort_depth = 0;
@@ -370,11 +222,13 @@ impl<'debuginfo> UnitInfo<'debuginfo> {
                                 match abstract_origin.value() {
                                     gimli::AttributeValue::UnitRef(unit_ref) => {
                                         if let Ok(abstract_die) = self.unit.entry(unit_ref) {
-                                            if let Some(mut die) = FunctionDie::new_inlined(
-                                                current.clone(),
-                                                abstract_die.clone(),
-                                                self,
-                                            ) {
+                                            if let Some(mut die) =
+                                                function_die::FunctionDie::new_inlined(
+                                                    current.clone(),
+                                                    abstract_die.clone(),
+                                                    self,
+                                                )
+                                            {
                                                 die.low_pc = ranges.begin;
                                                 die.high_pc = ranges.end;
